requirements.txt
wheel
requests
zmq
simplejson
gunicorn
flask
flask-wtf
flask-sqlalchemy
flask-migrate
flask-login
python-dotenv